Check if headscale node is alive before trying to contact traefik
In case the server is truly online, only then do you add the domains to
the list, otherwise drop them.

This should be handled a bit smarter, I think.
Perhaps maybe with a private database that keeps track of who owns what
domain and drop them only when the domain comes back online?

Anyway this will do for now.